Add `regexp/no-empty-capturing-group` rule that same `regexp/no-assertion-capturing-group` rule
